Move read/write_geometries to geometry module
